backend/requirements.txt
Flask==2.3.0
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-socketio==5.9.0
python-engineio==4.7.0
eventlet==0.33.3
python-dotenv==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.9
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
bcrypt==4.1.2
PyJWT==2.8.0